roachtest: add upgrade/foreign-key test
This PR adds a test that performs a rolling upgrade of a cluster from the
previous version to the current version, and adds and drops a foreign key at
every step of the rolling upgrade, concurrently with the TPCC load generator
running. It's primarily intended to test the forward- and
backward-compatibility of the changes to the foreign key representation on
table descriptors in 19.2 and 20.1.

Release note: None