jobs: pending jobs should be adoptable
It's possible that a job is created (so in state pending) but
due to a failure before the job is marked as started it can
be left in pending state. These jobs should be adoptable by
other nodes.

Touches #43375.

Release note (bug fix): this bug may cause jobs to be left indefinitely
in pending state and never run.